requirements.txt
Django==5.0.7
psycopg2==2.9.9
python-dotenv==1.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0
black
flake8
